added delay for creating app icons
on wayland - apps that use xwayland have delay between when the icon is available. In the old version I used restacked so it was no problem, but now I use the concrete window created event for better performance and efficiency, but it looks like wayland isn't about that. The current delay is 200ms and it works for me but I will most likely add it as an option as it can differ from one machine to another - like really slow ones taking even longer for the icon to become available on xwayland.