monitors changed fix
when monitors changed fires removed the line from renderer's destroy() that reset main monitor to what it was the last time render initialized. When monitors changed fires it changes the main monitor on its own. Bringing it back to the saved value has the potential to be incorrect which is guaranteed to break the extension.